Add AnsiConsole to CommandLineUtils project
